Flujo de Venta

+++ Domain/Entities/Item.cs
@@ -6,7 +6,7 @@
         public Guid VentaId { get; set; }
         public int ProductoId { get; set; }
         public int Cantidad { get; set; }
-        public decimal PrecioTotalItem { get; set; }
+        public int PrecioTotalItem { get; set; }
 
         public Venta Venta { get; set; }
         public Producto Producto { get; set; }
